move SelectionSource into module
